Move functions main() to direktiv main().
