+++ cmd/direktiv/main.go
@@ -5,6 +5,7 @@ import (
 	"github.com/direktiv/direktiv/cmd/api"
 	"github.com/direktiv/direktiv/cmd/flow"
 	flow_dbinit "github.com/direktiv/direktiv/cmd/flow-dbinit"
+	"github.com/direktiv/direktiv/cmd/functions"
 	"github.com/direktiv/direktiv/cmd/secrets"
 	"github.com/direktiv/direktiv/cmd/sidecar"
 	"log"
@@ -25,6 +26,8 @@ func main() {
 		flow.RunApplication()
 	case "flow_dbinit":
 		flow_dbinit.RunApplication()
+	case "functions":
+		functions.RunApplication()
 	case "":
 		log.Fatalf("error: empty DIREKTIV_APP environment variable.\n")
 	default:
